# docs/api/eino-workbench.openapi.json
{
  "openapi": "3.1.2",
  "info": {"title": "Eino Workbench M1 API", "version": "2.0.0-m1"},
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "getWorkbenchHealth",
        "responses": {"200": {"description": "Process is alive", "content": {"application/json": {"schema": {"type": "object", "required": ["status"], "properties": {"status": {"const": "ok"}}, "additionalProperties": false}}}}}
      }
    },
    "/readyz": {
      "get": {
        "operationId": "getWorkbenchReadiness",
        "responses": {"200": {"description": "M1 dependencies are ready", "content": {"application/json": {"schema": {"type": "object", "required": ["status"], "properties": {"status": {"const": "ready"}}, "additionalProperties": false}}}}}
      }
    },
    "/api/workspaces/{workspace_id}/views/current": {
      "get": {
        "operationId": "getCurrentWorkbenchView",
        "parameters": [{"name": "workspace_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}],
        "responses": {"200": {"description": "Current safe view", "content": {"application/json": {"schema": {"$ref": "../schemas/eino_workbench_view.v2.schema.json"}}}}}
      }
    },
    "/api/workspaces/{workspace_id}/messages": {
      "post": {
        "operationId": "postWorkbenchMessage",
        "parameters": [{"name": "workspace_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["conversation_id", "actor_id", "content"], "properties": {"conversation_id": {"type": "string", "minLength": 1}, "actor_id": {"type": "string", "minLength": 1}, "content": {"type": "string", "minLength": 1}}, "additionalProperties": false}}}},
        "responses": {"200": {"description": "Completed safe view", "content": {"application/json": {"schema": {"$ref": "../schemas/eino_workbench_view.v2.schema.json"}}}}}
      }
    },
    "/api/workspaces/{workspace_id}/runs/{run_id}": {
      "get": {
        "operationId": "getWorkbenchRunSnapshot",
        "parameters": [{"name": "workspace_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}, {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}],
        "responses": {"200": {"description": "Persisted safe view", "content": {"application/json": {"schema": {"$ref": "../schemas/eino_workbench_view.v2.schema.json"}}}}}
      }
    },
    "/api/workspaces/{workspace_id}/runs/{run_id}/stream": {
      "get": {
        "operationId": "streamWorkbenchRun",
        "parameters": [{"name": "workspace_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}, {"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1}}],
        "responses": {"200": {"description": "Safe SSE projection", "content": {"text/event-stream": {"schema": {"$ref": "../schemas/eino_workbench_stream_event.v2.schema.json"}}}}}
      }
    }
  }
}